+++ Sources/VCore/Extensions/SwiftUI/Color/Color.Dynamic.swift
@@ -11,7 +11,7 @@ import SwiftUI
 @available(watchOS, unavailable) // Light/dark mode and `UIColor(dynamicProvider:)` API do not exist. OS selects light color from the assets catalogue.
 @available(visionOS, unavailable) // Light/dark mode do not exist. OS selects dark color from the assets catalogue.
 extension Color {
-    /// Creates `UIColor` that generates it's color data dynamically.
+    /// Creates `Color` that generates it's color data dynamically.
     ///
     ///     let accentColor: Color = .dynamic(light: ..., dark: ...)
     ///